Create exfat_videorecovery.cpp
This will take in a disk image (created via dd) and will use the second volume boot record to reconstruct any .mp4 files left on the volume.  It hops through the FAT table and produces files from both contiguous and non-contiguous cluster maps.

This was useful for recovering videos from a GoPro Hero4 Silver SD Card that became unreadable by a computer.  YMMV, you may need to be a programmer yourself to make use of this.